requirements.txt
streamlit==1.29.0
sqlalchemy==2.0.23
alembic==1.13.1
pandas==2.1.4
python-dotenv==1.0.0
streamlit-authenticator==0.2.3
bcrypt==4.1.2
requests==2.31.0
python-dateutil==2.8.2
openpyxl==3.1.2
reportlab==4.0.7
pytz==2023.3